classify logical index scan into a separate file
